requirement.txt
# Core Spotify API interaction
spotipy
python-dotenv

# GUI library (for the modern macOS aesthetic)
customtkinter

# Data processing
Pillow
requests

# Optional: Required for YouTube URL input functionality
yt-dlp